Move 'tasks.py' inside 'public'
