+++ app.py
@@ -6,7 +6,7 @@ from flask_migrate import Migrate
 from portfolio_builder import create_app, db
 from portfolio_builder.auth.models import User
 from portfolio_builder.public.models import Security, Price, Watchlist, WatchlistItem
-from portfolio_builder.tasks import load_securities, load_prices
+from portfolio_builder.public.tasks import load_securities, load_prices
 
 
 app = create_app(settings_name='production')
